+++ app/controllers/items_controller.rb
@@ -1,16 +1,4 @@
 class ItemsController < ApplicationController
-  
-  def search
-    @keyword = params[:keyword]
-    @items = []
-    if @keyword.present?
-      @results = RakutenWebService::Ichiba::Product.search(keyword: @keyword, genreId: 101975)
-      @results.each do |result|
-        item = Item.new(read(result))
-        @items << item
-      end
-    end
-  end
 
   def index
   end
@@ -19,17 +7,17 @@ class ItemsController < ApplicationController
   end
 
   def create
-    
+    @item = Item.new(item_params)
+    if @item.save
+      redirect_to items_path, success: t('.success')
+    else
+      render :search
+    end
   end
 
   private
 
-  def read(result)
-    name = result["productName"]
-    url = result["productUrlPC"]
-    image_url = result["mediumImageUrl"]
-    maker = result["makerName"]
-    category = result["genreName"]
-    {name: name, url: url, image_url: image_url, maker: maker, category: category}
+  def item_params
+    params.require(:item).premit(:name, :url, :image_url, :maker, :category)
   end
 end
