+++ app/views/search_items/_item.html.erb
@@ -8,7 +8,6 @@
       <%= item.maker %></br>
       <%= item.name[0..20] ... %>
         <div class="menu">
-          <%= button_to "登録", action: "create", confirm: "登録しますか" %>
           <%= link_to "楽天でみる", item.url %>
         </div>
     </div>
